+++ lib/src/extract_path_pieces.dart
@@ -1,4 +1,6 @@
 List<String> extractPieces(String path) {
-  return path == '/' ? [] : path.split('/')
+  List<String> result = path == '/' ? [] : path.split('/')
     ..removeWhere((piece) => piece.isEmpty);
+
+  return result;
 }